Do not try to load recipes for a simulated environment

+++ pkg/recipes/engine/engine.go
@@ -28,6 +28,7 @@ import (
 	recipedriver "github.com/radius-project/radius/pkg/recipes/driver"
 	"github.com/radius-project/radius/pkg/recipes/util"
 	rpv1 "github.com/radius-project/radius/pkg/rp/v1"
+	"github.com/radius-project/radius/pkg/ucp/ucplog"
 )
 
 // NewEngine creates a new Engine to deploy recipe.
@@ -73,14 +74,22 @@ func (e *engine) Execute(ctx context.Context, opts ExecuteOptions) (*recipes.Rec
 // executeCore function is the core logic of the Execute function.
 // Any changes to the core logic of the Execute function should be made here.
 func (e *engine) executeCore(ctx context.Context, recipe recipes.ResourceMetadata, prevState []string) (*recipes.RecipeOutput, *recipes.EnvironmentDefinition, error) {
-	definition, driver, err := e.getDriver(ctx, recipe)
-	if err != nil {
-		return nil, nil, err
-	}
+	logger := ucplog.FromContextOrDiscard(ctx)
 
 	configuration, err := e.options.ConfigurationLoader.LoadConfiguration(ctx, recipe)
 	if err != nil {
-		return nil, definition, recipes.NewRecipeError(recipes.RecipeConfigurationFailure, err.Error(), util.RecipeSetupError, recipes.GetErrorDetails(err))
+		return nil, nil, recipes.NewRecipeError(recipes.RecipeConfigurationFailure, err.Error(), util.RecipeSetupError, recipes.GetErrorDetails(err))
+	}
+
+	// No need to try executing the recipe if it's a simulated environment.
+	if configuration.Simulated {
+		logger.Info("simulated environment enabled, skipping deployment")
+		return nil, nil, nil
+	}
+
+	definition, driver, err := e.getDriver(ctx, recipe)
+	if err != nil {
+		return nil, nil, err
 	}
 
 	// Retrieves the secret store id from the recipes configuration for the terraform module source of type git.
@@ -138,14 +147,18 @@ func (e *engine) Delete(ctx context.Context, opts DeleteOptions) error {
 // deleteCore function is the core logic of the Delete function.
 // Any changes to the core logic of the Delete function should be made here.
 func (e *engine) deleteCore(ctx context.Context, recipe recipes.ResourceMetadata, outputResources []rpv1.OutputResource) (*recipes.EnvironmentDefinition, error) {
-	definition, driver, err := e.getDriver(ctx, recipe)
+	configuration, err := e.options.ConfigurationLoader.LoadConfiguration(ctx, recipe)
 	if err != nil {
 		return nil, err
 	}
 
-	configuration, err := e.options.ConfigurationLoader.LoadConfiguration(ctx, recipe)
+	if configuration.Simulated {
+		return nil, nil
+	}
+
+	definition, driver, err := e.getDriver(ctx, recipe)
 	if err != nil {
-		return definition, err
+		return nil, err
 	}
 
 	err = driver.Delete(ctx, recipedriver.DeleteOptions{
